DAY-8 service updates

+++ src/main/java/com/kodilla/rental/service/ManufacturerDbService.java
@@ -2,6 +2,7 @@ package com.kodilla.rental.service;
 
 import com.kodilla.rental.domain.Manufacturer;
 import com.kodilla.rental.domain.dto.ManufacturerDto;
+import com.kodilla.rental.exception.alreadyExists.ManufacturerAlreadyExistException;
 import com.kodilla.rental.exception.notFound.ManufacturerNotFoundException;
 import com.kodilla.rental.mapper.ManufacturerMapper;
 import com.kodilla.rental.repository.ManufacturerRepository;
@@ -10,6 +11,8 @@ import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -28,21 +31,48 @@ public class ManufacturerDbService {
     }
 
     @Transactional
-    public ManufacturerDto createManufacturer(final ManufacturerDto manufacturerDto) {
-        Manufacturer manufacturer = manufacturerMapper.mapToManufacturer(manufacturerDto);
-        Manufacturer savedManufacturer = manufacturerRepository.save(manufacturer);
-        return manufacturerMapper.mapToManufacturerDto(savedManufacturer);
+    public ManufacturerDto createManufacturer(final ManufacturerDto manufacturerDto) throws ManufacturerAlreadyExistException {
+
+        List<String> allManufacturersNames = getAllManufacturers().stream()
+                .map(ManufacturerDto::getName)
+                .collect(Collectors.toList());
+
+        if (!allManufacturersNames.contains(manufacturerDto.getName())) {
+            Manufacturer manufacturer = manufacturerMapper.mapToManufacturer(manufacturerDto);
+            Manufacturer savedManufacturer = manufacturerRepository.save(manufacturer);
+            return manufacturerMapper.mapToManufacturerDto(savedManufacturer);
+        } else {
+            throw new ManufacturerAlreadyExistException();
+        }
     }
 
     @Transactional
-    public ManufacturerDto updateManufacturer(final ManufacturerDto manufacturerDto) {
-        Manufacturer manufacturer = manufacturerMapper.mapToManufacturer(manufacturerDto);
-        Manufacturer savedManufacturer = manufacturerRepository.save(manufacturer);
-        return manufacturerMapper.mapToManufacturerDto(savedManufacturer);
+    public ManufacturerDto updateManufacturer(final ManufacturerDto manufacturerDto) throws ManufacturerAlreadyExistException, ManufacturerNotFoundException{
+
+        List<String> allManufacturersNames = getAllManufacturers().stream()
+                .map(ManufacturerDto::getName)
+                .collect(Collectors.toList());
+
+        if (!manufacturerRepository.existsById(manufacturerDto.getManufacturerId())) {
+            throw new ManufacturerNotFoundException(manufacturerDto.getManufacturerId());
+        } else if (allManufacturersNames.contains(manufacturerDto.getName())) {
+            throw new ManufacturerAlreadyExistException();
+        } else {
+            Manufacturer manufacturer = manufacturerMapper.mapToManufacturer(manufacturerDto);
+            Manufacturer savedManufacturer = manufacturerRepository.save(manufacturer);
+            return manufacturerMapper.mapToManufacturerDto(savedManufacturer);
+        }
     }
 
     @Transactional
-    public void deleteManufacturer(final long manufacturerId) {
-        manufacturerRepository.deleteById(manufacturerId);
+    public void deleteManufacturer(final long manufacturerId) throws ManufacturerNotFoundException {
+
+        Optional<Manufacturer> manufacturer = manufacturerRepository.findById(manufacturerId);
+
+        if (manufacturer.isPresent()) {
+            manufacturerRepository.deleteById(manufacturerId);
+        } else {
+            throw new ManufacturerNotFoundException(manufacturerId);
+        }
     }
 }
